Rename test files to distinguish tx body and tx
